SQL Server Transport multi instance sample with Bridge (#5727)
