+++ samples/bridge/sql-multi-instance/Bridge_1/Helpers/SqlHelper.cs
