+++ samples/bridge/sql-multi-instance/Bridge_1/Receiver/OrderHandler.cs
@@ -1,0 +1,24 @@
+﻿using System.Threading.Tasks;
+using Messages;
+using NServiceBus;
+using NServiceBus.Logging;
+
+public class OrderHandler :
+    IHandleMessages<ClientOrder>
+{
+    static ILog log = LogManager.GetLogger<OrderHandler>();
+
+    #region Reply
+
+    public Task Handle(ClientOrder message, IMessageHandlerContext context)
+    {
+        log.Info($"Handling ClientOrder with ID {message.OrderId}");
+        var clientOrderAccepted = new ClientOrderResponse
+        {
+            OrderId = message.OrderId
+        };
+        return context.Reply(clientOrderAccepted);
+    }
+
+    #endregion
+}
