+++ samples/bridge/sql-multi-instance/Bridge_1/Receiver/Program.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Threading.Tasks;
+using NServiceBus;
+#pragma warning disable 618
+
+class Program
+{
+    const string ConnectionString = @"Data Source=.\SqlExpress;Database=NsbSamplesSqlMultiInstanceReceiver;Integrated Security=True;Max Pool Size=100";
+
+    static async Task Main()
+    {
+        Console.Title = "Samples.SqlServer.MultiInstanceReceiver";
+
+        #region ReceiverConfiguration
+
+        var endpointConfiguration = new EndpointConfiguration("Samples.SqlServer.MultiInstanceReceiver");
+        var transport = endpointConfiguration.UseTransport<SqlServerTransport>();
+        transport.ConnectionString(ConnectionString);
+        endpointConfiguration.EnableInstallers();
+
+        #endregion
+
+        SqlHelper.EnsureDatabaseExists(ConnectionString);
+
+        var endpointInstance = await Endpoint.Start(endpointConfiguration)
+            .ConfigureAwait(false);
+
+        Console.WriteLine("Press any key to exit");
+        Console.WriteLine("Waiting for Order messages from the Sender");
+        Console.ReadKey();
+        await endpointInstance.Stop()
+            .ConfigureAwait(false);
+    }
+}
