+++ samples/bridge/sql-multi-instance/Bridge_1/Shared/ClientOrder.cs
