+++ samples/bridge/sql-multi-instance/Bridge_1/Shared/ClientOrderResponse.cs
@@ -1,0 +1,11 @@
+﻿namespace Messages
+{
+    using System;
+    using NServiceBus;
+
+    public class ClientOrderResponse :
+        IMessage
+    {
+        public Guid OrderId { get; set; }
+    }
+}
