--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Bridge/Bridge.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <OutputType>Exe</OutputType>
-    <LangVersion>7.3</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus.SqlServer" Version="4.*" />
-    <PackageReference Include="NServiceBus.Router" Version="3.0.0-*" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Shared\Shared.csproj" />
-  </ItemGroup>
-</Project>
