--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/SQLServerMultiInstance.sln
@@ -1,52 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
-MinimumVisualStudioVersion = 15.0.26730.12
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "Shared\Shared.csproj", "{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Receiver.V3", "Receiver.V3\Receiver.V3.csproj", "{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sender.V3", "Sender.V3\Sender.V3.csproj", "{E1B21EF8-51A7-43BF-ACD0-653898075D2A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Receiver.V4", "Receiver.V4\Receiver.V4.csproj", "{49672AB4-75E1-4149-9A82-2EB0A102D5CF}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sender.V4", "Sender.V4\Sender.V4.csproj", "{D509E5B9-0E96-4318-9D40-D30238D7596F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "V3", "V3", "{55CD2DFC-BC89-4916-AEAA-A373DE30A872}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "V4", "V4", "{83711F32-80F3-4964-BA90-EE6D5ABF586F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bridge", "Bridge\Bridge.csproj", "{8F813F01-9D61-4C0C-A2D5-ABA68B7C0B8E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1B21EF8-51A7-43BF-ACD0-653898075D2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1B21EF8-51A7-43BF-ACD0-653898075D2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{49672AB4-75E1-4149-9A82-2EB0A102D5CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{49672AB4-75E1-4149-9A82-2EB0A102D5CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D509E5B9-0E96-4318-9D40-D30238D7596F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D509E5B9-0E96-4318-9D40-D30238D7596F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8F813F01-9D61-4C0C-A2D5-ABA68B7C0B8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8F813F01-9D61-4C0C-A2D5-ABA68B7C0B8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212} = {55CD2DFC-BC89-4916-AEAA-A373DE30A872}
-		{E1B21EF8-51A7-43BF-ACD0-653898075D2A} = {55CD2DFC-BC89-4916-AEAA-A373DE30A872}
-		{49672AB4-75E1-4149-9A82-2EB0A102D5CF} = {83711F32-80F3-4964-BA90-EE6D5ABF586F}
-		{D509E5B9-0E96-4318-9D40-D30238D7596F} = {83711F32-80F3-4964-BA90-EE6D5ABF586F}
-		{8F813F01-9D61-4C0C-A2D5-ABA68B7C0B8E} = {83711F32-80F3-4964-BA90-EE6D5ABF586F}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E6DA9E93-1AB4-41D3-ACEE-52F370D07637}
-	EndGlobalSection
-EndGlobal
