--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Sender.V4/OrderAcceptedHandler.cs
@@ -1,16 +1,0 @@
-﻿using System.Threading.Tasks;
-using Messages;
-using NServiceBus;
-using NServiceBus.Logging;
-
-public class OrderAcceptedHandler :
-    IHandleMessages<ClientOrderAccepted>
-{
-    static ILog log = LogManager.GetLogger<OrderAcceptedHandler>();
-
-    public Task Handle(ClientOrderAccepted message, IMessageHandlerContext context)
-    {
-        log.Info($"Received ClientOrderAccepted for ID {message.OrderId}");
-        return Task.CompletedTask;
-    }
-}
