--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Sender.V4/Sender.V4.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <OutputType>Exe</OutputType>
-    <LangVersion>7.3</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Shared\Shared.csproj" />
-    <PackageReference Include="NServiceBus" Version="7.*" />
-    <PackageReference Include="NServiceBus.SqlServer" Version="4.*" />
-    <PackageReference Include="NServiceBus.Bridge.Connector" Version="2.0.0-*" />
-  </ItemGroup>
-</Project>
