--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Shared/ClientOrder.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Messages
-{
-
-    public class ClientOrder
-    {
-        public Guid OrderId { get; set; }
-    }
-}
