--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Shared/ClientOrderAccepted.cs
@@ -1,9 +1,0 @@
-﻿namespace Messages
-{
-    using System;
-
-    public class ClientOrderAccepted
-    {
-        public Guid OrderId { get; set; }
-    }
-}
