--- samples/sqltransport/multi-instance-migration/SqlTransportLegacySystemClient_4/Shared/Shared.csproj
@@ -1,6 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <LangVersion>7.3</LangVersion>
-  </PropertyGroup>
-</Project>
