--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/Receiver/Program.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using NServiceBus;
-using NServiceBus.Transport.SQLServer;
-#pragma warning disable 618
-
-class Program
-{
-    static async Task Main()
-    {
-        Console.Title = "Samples.SqlServer.MultiInstanceReceiver";
-
-        #region ReceiverConfiguration
-
-        var endpointConfiguration = new EndpointConfiguration("Samples.SqlServer.MultiInstanceReceiver");
-        var transport = endpointConfiguration.UseTransport<SqlServerTransport>();
-        transport.EnableLegacyMultiInstanceMode(ConnectionProvider.GetConnection);
-        endpointConfiguration.UsePersistence<InMemoryPersistence>();
-        endpointConfiguration.SendFailedMessagesTo("error");
-        endpointConfiguration.EnableInstallers();
-
-        #endregion
-
-        transport.UseNativeDelayedDelivery().DisableTimeoutManagerCompatibility();
-
-        SqlHelper.EnsureDatabaseExists(ConnectionProvider.DefaultConnectionString);
-
-        var endpointInstance = await Endpoint.Start(endpointConfiguration)
-            .ConfigureAwait(false);
-
-        Console.WriteLine("Press any key to exit");
-        Console.WriteLine("Waiting for Order messages from the Sender");
-        Console.ReadKey();
-        await endpointInstance.Stop()
-            .ConfigureAwait(false);
-    }
-}
