--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/SQLServerMultiInstance.sln
@@ -1,29 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
-MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "Shared\Shared.csproj", "{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Receiver", "Receiver\Receiver.csproj", "{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sender", "Sender\Sender.csproj", "{E1B21EF8-51A7-43BF-ACD0-653898075D2A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50E54FA-4AE8-4018-BBA4-B8ED7748F214}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED0EC3CA-2B73-4E46-8CCA-DD390DF7A212}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1B21EF8-51A7-43BF-ACD0-653898075D2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1B21EF8-51A7-43BF-ACD0-653898075D2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E6DA9E93-1AB4-41D3-ACEE-52F370D07637}
-	EndGlobalSection
-EndGlobal
