--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/Shared/ClientOrderAccepted.cs
@@ -1,11 +1,0 @@
-﻿namespace Messages
-{
-    using System;
-    using NServiceBus;
-
-    public class ClientOrderAccepted :
-        IMessage
-    {
-        public Guid OrderId { get; set; }
-    }
-}
