--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/Shared/OrderReceived.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using NServiceBus;
-
-namespace Messages
-{
-    public class OrderReceived :
-        IEvent
-    {
-        public Guid OrderId { get; set; }
-    }
-}
