--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/Shared/Shared.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <LangVersion>7.3</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="6.*" />
-  </ItemGroup>
-</Project>
