--- samples/sqltransport/multi-instance/SqlTransportLegacySystemClient_3/Shared/SqlHelper.cs
@@ -1,52 +1,0 @@
-﻿using System.Data.SqlClient;
-
-public static class SqlHelper
-{
-    public static void ExecuteSql(string connectionString, string sql)
-    {
-        EnsureDatabaseExists(connectionString);
-
-        using (var connection = new SqlConnection(connectionString))
-        {
-            connection.Open();
-
-            using (var command = connection.CreateCommand())
-            {
-                command.CommandText = sql;
-                command.ExecuteNonQuery();
-            }
-        }
-    }
-
-    public static void CreateSchema(string connectionString, string schema)
-    {
-        var sql = $@"
-if not exists (select  *
-               from    sys.schemas
-               where   name = N'{schema}')
-    exec('create schema {schema}');";
-        ExecuteSql(connectionString, sql);
-    }
-
-    public static void EnsureDatabaseExists(string connectionString)
-    {
-        var builder = new SqlConnectionStringBuilder(connectionString);
-        var database = builder.InitialCatalog;
-
-        var masterConnection = connectionString.Replace(builder.InitialCatalog, "master");
-
-        using (var connection = new SqlConnection(masterConnection))
-        {
-            connection.Open();
-
-            using (var command = connection.CreateCommand())
-            {
-                command.CommandText = $@"
-if(db_id('{database}') is null)
-    create database [{database}]
-";
-                command.ExecuteNonQuery();
-            }
-        }
-    }
-}
