--- samples/sqltransport/multi-instance/sample_passconnection_sqltransportlegacysystemclient_[,3.0).partial.md
@@ -1,1 +1,0 @@
-NOTE: In versions 2 and below it is required to pass connection string for the currently configured endpoint either using the `ConnectionString()` method or in the `app.config` file.
